Refactored pokemon table.
Extracted summary with the search logic into it's own component. Removed
JSON server as deemed uncessary. Updated Documentation.
