+++ README.md
@@ -1,11 +1,12 @@
 # Product
-The product is a simple pokemon data table with pagination that displays most of pokemons with their appropriate attributes (name, type, attack, defense..). In addition to the ability of search bu name and power thershold. 
+
+The product is a simple pokemon data table with pagination that displays most of pokemons with their appropriate attributes (name, type, attack, defense..). In addition to the ability of search bu name and power thershold.
 
 Screenshots
 
-# Technologies: 
-React, Scss, Axios, npm, Vite, Json-Sever, concurrently, Material UI.
+# Technologies:
 
+React, Scss, Axios, npm, Vite, Json-Sever, concurrently, Material UI.
 
 Backend API
 No backend, just local data handled by Json-Server.
@@ -16,15 +17,15 @@ No backend, just local data handled by Json-Server.
 │ ─── .github # CI/CD workflow
 │ ─── public
 │ ├─── pokemon.json # local pokemon data
-│ ├─── vite.svg # Placeholder 
+│ ├─── vite.svg # Placeholder
 │ ───src # Main application folder
-│ ├───assets # Application assets: Pictures, Icons. 
+│ ├───assets # Application assets: Pictures, Icons.
 │ ├───components # UI Components used
-│ │ ├───DataTable # element for rendering data as a table
+│ │ ├─── DataTable # element for rendering data as a table
 │ │ │ └── DataTable.jsx # DataTable component
 │ │ │ └── DataTable.scss # DataTable stylesheeet
-│ │ ├───Pagination # element for pagination creeated with MUI
-│ │ │ └── Pagination.jsx # Pagination component
+│ │ ├─── SummarySectio # element for pagination creeated with MUI
+│ │ │ └── SummarySection.jsx # Pagination component
 │ │ ├───SearchBar # element for rendering search input
 │ │ │ └── SearchBar.jsx # SearchBar component
 │ │ │ └── SearchBar.scss # SearchBar stylesheeet
@@ -33,16 +34,16 @@ No backend, just local data handled by Json-Server.
 │ │ ├───useDebounce.js # hook for debouncing search for better optimization
 │ ├───services # Data manipulation
 │ ├───pages
-│ │ ├─── PokemonTable # View for pokemon table 
+│ │ ├─── PokemonTable # View for pokemon table
 │ │ │ └── PokemonTable.jsx # PokemonTable component
 │ │ │ └── PokemonTable.scss # PokemonTable stylesheeet
 │ └─── index.js # file for facillitationg imports
-│ ├───services 
+│ ├───services
 │ │ ├─── pokemon.json # Pokemon data fetching and manipulation
 │ ├───styles
 │ │ ├─── _base.scss # CSS reset file
 │ │ ├─── _variable.scss # CSS variables
-│ │ ├─── index.scss # common css classes 
+│ │ ├─── index.scss # common css classes
 │ ├───utils
 │ │ ├─── helper.js # Useful manipulation functions
 │ ├───app.jsx # App component
@@ -55,3 +56,9 @@ No backend, just local data handled by Json-Server.
 │ ───Readme.md # Documentation
 │ ─── vite.config.js # Vite configuration
 ```
+
+# Remarks
+
+1. The project is supported with a simple deployment pipline. you can find the project in github pages at this link [https://fadibenz.github.io/Legal-Doctrine-Frontend/]
+2. I used a debounce function on the search to limit filtering times, the search might seem slow but it's intended.
+3. the search is by 'name' and the power is calccultated by summing all attributes of the pokemon.
