+++ package.json
@@ -4,8 +4,7 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "start:server": "json-server --watch public/pokemon.json --port 3001",
-    "dev": "concurrently \"vite\" \"npm run start:server\"",
+    "dev": "vite",
     "build": "vite build",
     "eslint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
@@ -17,7 +16,6 @@
     "@mui/icons-material": "^5.14.18",
     "@mui/material": "^5.14.18",
     "axios": "^1.6.2",
-    "concurrently": "^8.2.2",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
