+++ src/components/index.js
@@ -1,5 +1,4 @@
 import SearchBar from "./SearchBar/SearchBar";
 import DataTable from "./DataTable/DataTable";
-import Pagination from "./Pagination/Pagination";
 
-export { SearchBar, DataTable, Pagination };
+export { SearchBar, DataTable };
