+++ src/pages/PokemonTable/PokemonTable.jsx
@@ -1,16 +1,11 @@
 import { useEffect, useState } from "react";
 import { getData } from "../../services/pokemon";
-import {
-  calculatePower,
-  sliceData,
-  handlePowerChange,
-} from "../../utils/helper";
-import { useDebounce } from "../../hooks/useDebounce";
-import { useField } from "../../hooks/useField";
-import { Pagination, DataTable, SearchBar } from "../../components";
-import { TbHeartPlus } from "react-icons/tb";
-import { TfiSearch } from "react-icons/tfi";
+import { sliceData, handlePowerChange } from "../../utils/helper";
+import { DataTable } from "../../components";
+import { TablePagination } from "@mui/material";
+
 import "./PokemonTable.scss";
+import SummarySection from "../../components/SummarySection/SummarySection";
 
 const tableHead = [
   "ID",
@@ -42,11 +37,6 @@ export default function PokemonTable() {
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
 
-  const searchInput = useField("");
-  const powerTH = useField("");
-  const debounceSearch = useDebounce(searchInput.value, 300);
-  const debouncePower = useDebounce(powerTH.value, 300);
-
   const [power, setPower] = useState({
     min: 0,
     max: 0,
@@ -61,6 +51,7 @@ export default function PokemonTable() {
     setPage(0);
   };
 
+  // fetching data
   useEffect(() => {
     const fetchData = async () => {
       try {
@@ -76,6 +67,7 @@ export default function PokemonTable() {
     fetchData();
   }, []);
 
+  // slicing data for pagination
   useEffect(() => {
     const newData = sliceData(changeData.filteredData, page, rowsPerPage);
     setChangeData({ ...changeData, slicedData: newData });
@@ -83,42 +75,24 @@ export default function PokemonTable() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [page, rowsPerPage, changeData.filteredData]);
 
-  useEffect(() => {
-    const searchData = async () => {
-      let newData = data.data;
-      if (debounceSearch !== "") {
-        newData = newData.filter((pokemon) =>
-          pokemon.name.toUpperCase().includes(debounceSearch.toUpperCase())
-        );
-      }
-      if (debouncePower !== "") {
-        newData = newData.filter(
-          (pokemon) => calculatePower(pokemon) >= debouncePower
-        );
-      }
-      setChangeData({
-        ...changeData,
-        filteredData: newData,
-        count: newData.length,
-      });
-    };
-    searchData();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [debounceSearch, debouncePower, data]);
-
   return (
     <section className='Container'>
-      <SummarySection />
-      <article>
-        <DataTable data={changeData.slicedData} tableHead={tableHead} />
-        <Pagination
-          count={changeData.count}
-          page={page}
-          rowsPerPage={rowsPerPage}
-          handleChangePage={handleChangePage}
-          handleChangeRowsPerPage={handleChangeRowsPerPage}
-        />
-      </article>
+      <SummarySection
+        data={data.data}
+        setChangeData={setChangeData}
+        changeData={changeData}
+        power={power}
+      />
+      <DataTable data={changeData.slicedData} tableHead={tableHead} />
+      <TablePagination
+        component='div'
+        count={changeData.count}
+        page={page}
+        onPageChange={handleChangePage}
+        rowsPerPage={rowsPerPage}
+        onRowsPerPageChange={handleChangeRowsPerPage}
+        rowsPerPageOptions={[5, 10, 15]}
+      />
     </section>
   );
 }
