+++ src/services/pokemon.js
@@ -2,8 +2,9 @@ import axios from "axios";
 
 export const getData = async () => {
   try {
-    const baseURL = `http://localhost:3001/data`;
+    const baseURL = `/pokemon.json`;
     const response = await axios.get(baseURL);
+    console.log('response', response)
     return {
       data: response.data,
       count: response.data.length,
